feat: LINUX, open file on editor, setup using $EDITOR env var.
This feature use the EDITOR environment variable to open a editor
when enter is clicked meanwhile a file is selected in the file list.

I think it works in MacOS but I don't have a Mac to test and
confirm it so I'm not sure, you are free to do it if you want.